break add-course out into it's own file
